+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added Ability to delete notes
 - Fixed Spelling in README.md
 - Changed Formatting in Changeog
+- Recoded part of the "Create" function
 
 #### 0.1.0 - First Relase (7/5/2020)